@@ -5,6 +5,13 @@
 </svelte:head>
 <script>
   import "../app.css";
+  /**
+   * @typedef {Object} Props
+   * @property {import('svelte').Snippet} [children]
+   */
+
+  /** @type {Props} */
+  let { children } = $props();
 </script>
 
-<slot />
+{@render children?.()}
