@@ -22,4 +22,4 @@ export const hooks = {
 	reroute: (() => {})
 };
 
-export { default as root } from '../root.svelte';
+export { default as root } from '../root.js';
